install.js: Fail the postinstall script if compiling fails
If the pngquant binary does not work and the compile fails, npm should
fail the package installation, since it will not work. Previously it
would log an error, but return exit code 0 so npm would think the
package was installed correctly. This will make it easier for users to
figure out why pngquant isn't working.